Elaborate on repairs; more verbose collapsibles
- Talk more about the process of repairing Flexspirators. Drastic
repairs (for Flexspirators that have completely broken) are probably
unnecessary for nearly all users; painting over perforations with Flex
Seal Liquid is probably the most that any person would need.

- Make collapsible text box summaries more verbose, so that it isn't
necessary to click-to-expand to understand the point of the warning. If
a person prints out these instructions, but forgets to click-to-expand
the collapsibles, then the contents of the collapsible would be omitted
from the printout.

- Mention that new Flexspirators use a soft rubber weather stripping and
not vinyl weather stripping.

- Fix broken image showing how to smear petroleum jelly around the back
edge of the Filter Tray.

- Link to page history.

- Fix typos.